+++ airbyte-integrations/connectors/source-close-com/source_close_com/manifest.yaml
@@ -510,3 +510,31 @@ streams:
 check:
   stream_names:
     - "users"
+
+spec:
+  type: Spec
+  documentation_url: https://docs.airbyte.com/integrations/sources/close-com
+  connection_specification:
+    $schema: http://json-schema.org/draft-07/schema#
+    title: Close.com Spec
+    type: object
+    required:
+    - api_key
+    additionalProperties: true
+    properties:
+      api_key:
+        title: API Key
+        type: string
+        description: Close.com API key (usually starts with 'api_'; find yours <a href="https://app.close.com/settings/api/">here</a>).
+        airbyte_secret: true
+      start_date:
+        title: Replication Start Date
+        type: string
+        description: 'The start date to sync data; all data after this date will be
+          replicated. Leave blank to retrieve all the data available in the account.
+          Format: YYYY-MM-DD.'
+        examples:
+        - '2021-01-01'
+        default: '2021-01-01'
+        pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
+        format: date
