--- airbyte-integrations/connectors/source-close-com/source_close_com/source_lc.py
@@ -1,18 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-from airbyte_cdk.sources.declarative.yaml_declarative_source import YamlDeclarativeSource
-
-"""
-This file provides the necessary constructs to interpret a provided declarative YAML configuration file into
-source connector.
-
-WARNING: Do not modify this file.
-"""
-
-
-# Declarative Source
-class SourceCloseCom(YamlDeclarativeSource):
-    def __init__(self):
-        super().__init__(**{"path_to_yaml": "manifest.yaml"})
